refactor(sync): restructure firewall permission card into MVC architecture
Split the large firewall permission card component into separate files following MVC pattern:
- Created controller for state management and business logic
- Extracted helper for platform-specific operations
- Moved manual setup guide to dedicated component
- Updated main card component to use new architecture

This improves maintainability and separation of concerns in the sync feature.
